refactor: LocalFormatted~ 를 widget 패키지 외부로 이동
