+++ app/src/main/java/woowacourse/movie/view/LocalFormattedTime.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.widget
+package woowacourse.movie.view
 
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter
