+++ app/src/main/java/woowacourse/movie/view/widget/DateSpinner.kt
@@ -3,6 +3,7 @@ package woowacourse.movie.view.widget
 import android.os.Bundle
 import android.view.View
 import android.widget.AdapterView
+import woowacourse.movie.view.LocalFormattedDate
 import woowacourse.movie.view.MovieViewData
 
 class DateSpinner(val spinner: SaveStateSpinner) {
